+++ testing/auth_types.ts
@@ -73,7 +73,7 @@ export type Credentials =
 interface BaseOauth2RequestAccessTokenParams {
   client_id: string;
   client_secret: string;
-  resource?: string | string[];
+  resource?: string;
 }
 
 export interface OAuth2RequestAccessTokenParams extends BaseOauth2RequestAccessTokenParams {